Start simplifying JuliaJITEventListener cruft [nfc]
The same interface used to be called two different ways depending on
the JIT implementation: via llvm::JITEventListener for use with the
old JIT APIs, as well as through our own ORCNotifyObjectEmitted()
callback for use with the new ORC-based JIT. Legacy JIT and MCJIT
support is long gone, so this removes the unused
llvm::JITEventListener interface.

Since there was a single instance held in a static global in
debuginfo.cpp anyway, I also removed CreateJuliaJITEventListener(),
etc. in favour of a single static global instance – this of course
still isn't exactly clean lifetime management.